md_recommendation_system/requirements.txt
torch
torchmetrics
torchvision
scikit-learn
numpy
pandas
tqdm
PyYAML
pymongo
transformers
boto3
jieba